fix(python): close async affine loop ownership gaps
